requirements.txt
<<<<<<< HEAD
# Веб-фреймворк
streamlit==1.29.0

# Парсинг документов
pdfplumber==0.10.3
python-docx==1.1.0

# Локальная LLM через Ollama
ollama==0.1.6
langchain==0.0.340
langchain-community==0.0.20

# Векторная БД (опционально)
chromadb==0.4.18

# Вспомогательные
python-dotenv==1.0.0
pandas==2.1.3

# Для эмбеддингов (бесплатно, локально)
=======
# Веб-фреймворк
streamlit==1.29.0

# Парсинг документов
pdfplumber==0.10.3
python-docx==1.1.0

# Локальная LLM через Ollama
ollama==0.1.6
langchain==0.0.340
langchain-community==0.0.20

# Векторная БД (опционально)
chromadb==0.4.18

# Вспомогательные
python-dotenv==1.0.0
pandas==2.1.3

# Для эмбеддингов (бесплатно, локально)
>>>>>>> 5c3629e5d0a871d9a4a3e69489dfd8cd17e69293
sentence-transformers==2.2.2